requirements.txt
# Core
tensorflow>=2.15,<2.17
tensorflow-datasets>=4.9
opencv-python-headless>=4.8
Pillow>=10.0
scikit-image>=0.21
PyYAML>=6.0

# OCR engines
pytesseract>=0.3.10
paddleocr>=2.7
transformers>=4.36
torch>=2.1          # needed for TrOCR

# API
fastapi>=0.109
uvicorn[standard]>=0.25
python-multipart>=0.0.6

# Dashboard
streamlit>=1.30

# Post-processing
symspellpy>=6.7
reportlab>=4.0
PyMuPDF>=1.23
lxml>=5.0

# Data & training
albumentations>=1.3
matplotlib>=3.8
pandas>=2.1
numpy>=1.24
tqdm>=4.66
tensorboard>=2.15

# Evaluation
editdistance>=0.6
jiwer>=3.0

# Settings
pydantic-settings>=2.0

# Testing
pytest>=7.4
httpx>=0.25      # for testing FastAPI
